Refacto SDK as Singleton
